refactor, use new schema director
